APIv1-v3 fixes, APIv6 beta
