+++ AlienFX_SDK/AlienFX_SDK.vcxproj.filters
@@ -18,6 +18,9 @@
     <ClInclude Include="AlienFX_SDK.h">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="alienfx-controls.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="AlienFX_SDK.cpp">
